[LLVM 16][build] Update Docker image for building LLVM on Linux
CMake version in the current image is too old to build LLVM 16:
the image has CMake 3.5.1, while the build requires CMake 3.13.4
or higher.

This commit updates the Docker image from Ubuntu 16.04 to Ubuntu 20.04,
which features CMake 3.16.3.
It is important not to use a too recent version of image, because that
would limit the compatibility of the built LLVM with older OS versions.

^KTI-1673
^KT-49279

Co-authored-by: Svyatoslav Scherbina <svyatoslav.scherbina@jetbrains.com>