[LLVM 16] Fix libclang interop stub generator
To allow using libclang from Kotlin, we have auto-generated stubs
that are compiled with the bundled LLVM version.

Updating to LLVM 16 leads to using Clang 16 for stub compilation.
Clang 16 is more strict, so the stubs (=> the generator) should be
adjusted: an explicit cast to void* is necessary when calling `memcpy`.

^KT-49279