Add Compose compiler plugin to IDE plugin dependencies
This allows IDE to use the Compose compiler plugin from the KT compiler.

In addition, this commit reimplements the Compose compiler plugin
registrar based on the new API `CompilerPluginRegistrar` because K2 IDE
supports putting a compiler plugin to the allow-list only when the
registrar is a child class of `CompilerPluginRegistrar`.

KTIJ-30164