[LLVM 16] Workaround for old ld.gold and lldb bugs
Using the retain attribute hits a bug in the currently bundled version
of ld.gold, resulting in two separate sections being created, one for
symbols with retain and one for symbols without. Both these sections are
named .text. This isn't technically incorrect ELF, but it in turn hits a
bug in LLDB, where symbols in the second .text are incorrectly resolved,
with offsets relative to the first .text section. The upshot of all this
is that adding retain will make LLDB unable to lookup and call the
Konan_DebugXX functions required for the konan_lldb.py extension.

This fix removes the retain annotation from the debug functions, thus
moving them to the first .text section, where they get resolved
correctly in old lldb version. To keep the symbols alive, a dummy array
of pointers to those functions is annotated with retain.

Both of the bugs are fixed upstream in their respective projects, so
this workaround is only required until, e.g., the linker gets upgraded
or changed.

^KT-69206 Fixed