[FIR] Add a test for `FirAccidentalOverrideClashChecker`
This test is sensitive to the specified `MppCheckerKind.Platform`:
`Common` would erase `ACCIDENTAL_OVERRIDE_CLASH_BY_JVM_SIGNATURE`.
The diagnostic must be present, though, otherwise through the `expect`
class it would be possible to declare something that is incorrect
from the PoV of the `actual` class (can't be declared there).

^KT-68000