FIR: always copy property accessors for delegated origin
And also, keep metadata for property accessors when changing origin of
fake overrides to delegation. This metadata is used in the backend when
generating JVM signatures in the metadata, so this fixes KT-68997.

It also fixes simple non-generic cases of KT-63828. In case the
interface (delegation target) is generic, FirFakeOverrideGenerator only
gets members with origin SubstitutionOverride instead of Delegated.
This needs more investigation.

 #KT-63828
 #KT-63871 Fixed
 #KT-68997 Fixed