IR Inliner: Add test for references for stub local inline functions
Such functions are generated by CommonInlineCallableReferenceToLambdaPhase lowering. And they have local visibility which makes it difficult for the IR inliner to deduce on which exactly phase of double-inlining to inline them.

We need to teach the inliner to handle such references. And then remove the guard condition added to IrValidationAfterInliningOnlyPrivateFunctionsPhase.

^KT-69172
^KT-69470